add class javadocs
